feat : User에 있던 내 정보 조회와 프로필 사진 관련 기능들을 MyPage 기능으로 분리 및 테스트 코드 분리
