+++ src/main/java/com/shcho/shBlog/user/controller/UserController.java
@@ -1,14 +1,18 @@
 package com.shcho.shBlog.user.controller;
 
-import com.shcho.shBlog.auth.CustomUserDetails;
-import com.shcho.shBlog.user.dto.*;
+import com.shcho.shBlog.user.dto.UserSignInRequestDto;
+import com.shcho.shBlog.user.dto.UserSignInResponseDto;
+import com.shcho.shBlog.user.dto.UserSignUpRequestDto;
+import com.shcho.shBlog.user.dto.UserSignUpResponseDto;
 import com.shcho.shBlog.user.entity.User;
 import com.shcho.shBlog.user.service.UserService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/auth")
@@ -38,35 +42,6 @@ public class UserController {
         UserSignInResponseDto responseDto = UserSignInResponseDto.of(user, token);
 
         return ResponseEntity.ok(responseDto);
-    }
-
-    @GetMapping("/me")
-    public ResponseEntity<UserInfoResponseDto> getMyInfo(
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        User user = userDetails.getUser();
-        UserInfoResponseDto responseDto = UserInfoResponseDto.from(user);
-
-        return ResponseEntity.ok(responseDto);
-    }
-
-    @PatchMapping("/profile")
-    public ResponseEntity<String> updateProfile(
-            @AuthenticationPrincipal CustomUserDetails userDetails,
-            @RequestBody UserProfileRequestDto requestDto
-    ) {
-        Long userId = userDetails.getUser().getUserId();
-        userService.updateProfileImage(userId, requestDto.profileImageUrl());
-        return ResponseEntity.ok("프로필 이미지 등록 완료");
-    }
-
-    @DeleteMapping("/profile")
-    public ResponseEntity<String> deleteProfile(
-            @AuthenticationPrincipal CustomUserDetails userDetails
-    ) {
-        Long userId = userDetails.getUser().getUserId();
-        userService.deleteProfileImage(userId);
-        return ResponseEntity.ok("프로필 이미지 삭제 완료");
     }
 }
