+++ src/test/java/com/shcho/shBlog/user/service/UserServiceTest.java
@@ -30,9 +30,6 @@ class UserServiceTest {
     @Mock
     private PasswordEncoder passwordEncoder;
 
-    @Mock
-    private S3Service s3Service;
-
     @InjectMocks
     private UserService userService;
 
@@ -208,109 +205,5 @@ class UserServiceTest {
         return new UserSignInRequestDto(username, password);
     }
 
-    @Test
-    @DisplayName("프로필 사진 업데이트 성공")
-    void updateUserProfileImageSuccess() {
-        // given
-        Long userId = 1L;
-        String newImageUrl = "www.minio.com/new.jpg";
-        String oldImageUrl = "www.minio.com/old.jpg";
 
-        User user = User.builder()
-                .userId(userId)
-                .username("existsUsername")
-                .nickname("test")
-                .email("test@email.com")
-                .password("encodedPassword")
-                .profileImageUrl(oldImageUrl)
-                .role(USER)
-                .build();
-
-        when(userRepository.getReferenceById(userId)).thenReturn(user);
-
-        // when
-        userService.updateProfileImage(userId, newImageUrl);
-
-        // then
-        verify(s3Service, times(1)).deleteFileByUrl(oldImageUrl);
-        assertEquals(newImageUrl, user.getProfileImageUrl());
-
-    }
-
-    @Test
-    @DisplayName("프로필 사진 업데이트 성공 - 기존 이미지 없음")
-    void updateUserProfileImageWithoutOldImage() {
-        // given
-        Long userId = 1L;
-        String newImageUrl = "www.minio.com/new.jpg";
-
-        User user = User.builder()
-                .userId(userId)
-                .username("existsUsername")
-                .nickname("test")
-                .email("test@email.com")
-                .password("encodedPassword")
-                .profileImageUrl(null)
-                .role(USER)
-                .build();
-
-        when(userRepository.getReferenceById(userId)).thenReturn(user);
-
-        // when
-        userService.updateProfileImage(userId, newImageUrl);
-        verify(s3Service, never()).deleteFileByUrl(newImageUrl);
-    }
-
-    @Test
-    @DisplayName("프로필 사진 삭제 성공")
-    void deleteUserProfileImageSuccess() {
-        // given
-        Long userId = 1L;
-        String oldImageUrl = "www.minio.com/old.jpg";
-
-        User user = User.builder()
-                .userId(userId)
-                .username("existsUsername")
-                .nickname("test")
-                .email("test@email.com")
-                .password("encodedPassword")
-                .profileImageUrl(oldImageUrl)
-                .role(USER)
-                .build();
-
-        when(userRepository.getReferenceById(userId)).thenReturn(user);
-
-        // when
-        userService.deleteProfileImage(userId);
-
-        // then
-        verify(s3Service, times(1)).deleteFileByUrl(oldImageUrl);
-        assertNull(user.getProfileImageUrl());
-    }
-
-    @Test
-    @DisplayName("프로필 사진 삭제 - 기존 이미지 없음")
-    void deleteUserProfileImageWithoutOldImage() {
-        // given
-        Long userId = 1L;
-
-        User user = User.builder()
-                .userId(userId)
-                .username("existsUsername")
-                .nickname("test")
-                .email("test@email.com")
-                .password("encodedPassword")
-                .profileImageUrl(null) // 이미지 없음
-                .role(USER)
-                .build();
-
-        when(userRepository.getReferenceById(userId)).thenReturn(user);
-
-        // when
-        userService.deleteProfileImage(userId);
-
-        // then
-        verify(s3Service, never()).deleteFileByUrl(anyString());
-        assertNull(user.getProfileImageUrl());
-    }
 }
